fix: use BigIntegerField for DeletedAccount.account_id
Some WG account IDs exceed 2^31-1 (max 3,012,966,527 in the dataset).

+++ server/warships/models.py
@@ -287,7 +287,7 @@ class PlayerAchievementStat(models.Model):
 
 class DeletedAccount(models.Model):
     """Permanently blocklisted Wargaming account IDs (GDPR / account deletion)."""
-    account_id = models.IntegerField(unique=True)
+    account_id = models.BigIntegerField(unique=True)
     deleted_at = models.DateTimeField(auto_now_add=True)
 
     class Meta:
